BISERVER-7280 Updated get jobs to check if an user is Admin or not. Moved canSchedule to SchedulerResource
